requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
prophet
statsmodels
torch
pytorch-lightning
plotly
pyyaml
fastparquet
